requirements.txt
django-autocomplete-light
requests